jpa-hibernate/jpa-mapping/src/main/java/mapping/converter/LocalDateConverter.java
package mapping.converter;

import java.sql.Date;
import java.time.LocalDate;

import javax.persistence.*;

// TODO 05: Create a JPA 2.1 converter for java.time.LocalDate 

// TODO 05c: Mark this as a converter
public class LocalDateConverter implements AttributeConverter<LocalDate, Date> {

	@Override
	public Date convertToDatabaseColumn(LocalDate localDate) {
		// TODO 05a: Convert java.time.LocalDate to java.sql.Date
		// Hint: Use static java.sql.Date#valueOf(LocalDate) factory method
		// Run LocalDateConverterTests to see that it is working
		return null;
	}

	@Override
	public LocalDate convertToEntityAttribute(Date date) {
		// TODO 05b: Convert java.sql.Date to java.time.LocalDate
		// Hint: Use static java.sql.Date#toLocalDate() factory method
		// Run LocalDateConverterTests to see that it is working
		return null;
	}

}
